# src/main/resources/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/authors/{id}": {
      "get": {
        "tags": [
          "author-controller"
        ],
        "summary": "Retrieves an author by his ID.",
        "description": "Retrieves an author by his ID.",
        "operationId": "getAuthor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "the ID of the author to retrieve",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "withBooks",
            "in": "query",
            "description": "whether to include the author\u0027s books in the response",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "a ResponseEntity containing the GetAuthorResponse",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GetAuthorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Book": {
        "type": "object",
        "properties": {
          "bookId": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "publishedAt": {
            "type": "string"
          }
        }
      },
      "GetAuthorResponse": {
        "type": "object",
        "properties": {
          "exists": {
            "type": "boolean"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "Name": {
            "type": "string"
          },
          "books": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Book"
            }
          }
        }
      }
    }
  }
}